Annotations now adding correctly.
Polyglot was adding annotations using the OWLNamedEntity which
is sensible. However, add-annotation was passing this through
ensure-class which was crashing when polyglot was passing in a
property. So, add-annotation now expects a OWLNamedEntity and no
longer calls ensure-class. However, owlclass-explicit was calling
add-annotation (and several other methods) using the classname rather
than the OWLClass object (even though it has this available to it).
So this has been fixed as well.