Dispose of all reasoners when creating a new factory.
Previously, reasoners were being disposed when an ontology
was created, but not when the reasoner factory was changed.
This causes especially problems for reasoner_test.clj which
does this a lot. It was causing tests to be non-independent
of each other which was very confusing. It would probably
also cause a resource consumption issue.