IRI no longer stored in Ontology record, only in OWLOntology.
Previously, the IRI for each ontology was stored in the clojure
record. However, this is unnecessary as it is also available in the
OWLOntology object. This required a new function to get the IRI from
an ontology. get-current-iri has been deprecated as the same function
is now fulfilled by a no-args call to get-iri.

Several other fixes involved removing direct access to the Ontology
object and to use the get-iri method instead.