Removed Options from Ontology record.
The ontology record was carrying a ref to a hash which
was used to store options -- that is, anything that I want
to add to be carried around by the ontology. Options are now
stored externally in a map. This will allow me to remove to
record and just use an OWLOntology object directly.