Initial state of vreasoner-factory fixed.
For some reason, this was set to the empty list, which makes no sense, as in
use it carries a single reasoner factory. Now set to nil at start up. Added a
single test case, to at least check the exception, although it can't test the
real startup value, as tests do not run in the order they are written.