ensure-class now handles a closure arg arity > 0.
Previously when passed a closure of arity greater than one ensure-class
crashes with an ArityException when it calls the closure. However, the fact
and fact-not functions also return closures, but with arity = 1. The can
easily (but incorrectly) be used where a class might be expected, so a better
crash makes sense.

The ArityException is caught, and an IllegalArgumentException is rethrown.