owl-not overloaded for facts, is function added.
Previously, the fact-not function had to be called directly. owl-not
can now be used instead which delegates its functionality to fact-not.
This is consistent with object-not and data-not. And is function has been
added as a counterpart to owl-not for cleanlineness.

To achieve this, the old owl-not multimethod function has been renamed to
owl-not-one (and made private). A new function called owl-not has been added
which is simply overloaded on arity -- one argument calls the multimethod,
two arguments calls fact-not.

The is function is a simple alias.