VersionInfo now handled correctly.
VersionInfo now handled the same was as other ontology keywords, which
means that they work correctly. Keywords to the ontology function are
now checked (as with other functions) to see that a handler exists. This
was hiding a broken test case. Finally, vtawny-ontology now uses
defonce, otherwise a reload results in inconsistencies.

Closes #13