p now returns a record, making e is superfluous.
Originally, p returned a vector, with e being used to extract the entity
from this vector. However, the requirement to use e was a painful
imposition of implementation detail on the user. By moving p to return a
record and extending that record to Entityable, we no longer need to
explicitly unpack the record as owl.clj already does this using as-entity.