Pattern annotations added.
It is now possible to annotate a set of entities as belonging to a
pattern, distinguishable from other patterns, including instances of the
same pattern. There are some utility functions which return the pattern,
and other entities in that pattern.

The pattern data is stored as annotations, using a
OWLAnonymousIndividual to define each pattern. The individuals are
cached partly for speed, but mostly so that multiple evaluations of a
pattern are idempotent and do not add new individuals to each pattern.

The functions have also been used for value partition. The pattern
namespace is not reflection tested and can be fast use'd from the REPL.