iri-gen functions now passed an ontology.
iri-gen functions translate between tawny-names and ontology IRIs.
Currently, there are (effectively) two of these functions in tawny. The
"default" case is handled separately and uses the ontology as an
argument. The other, in obo.clj, is a "default" ontology function, so
uses an ontology but requires that it is set as the default.

We change this to explicitly pass an ontology function now. The
"default" case shows that it is necessary (to capture the ontology IRI
if for no other reason), and the obo.clj shows that it will probably be
handled silently anyway.