Enable niling of the reasoner factory.
Previously, the reasoner factory started in a nil state but could not be
set back to this value again. This has now been enabled through the
reasoner-factory method.

We no longer dispose of reasoners when changing factory, as this was
causing exceptions in both Jfact and Hermit. Finally, :hermit has been
re-enabled in testing.