Improve keyword parsing for object-property
Previously, keywords at which to split where explicitly specified. But this
gives poor error message when wrong keywords are given, so now, instead, we
identify keywords at which NOT to split. This also addresses an error where
incorrect keywords as a first argument were silently dropped.

Also update datatype-property, datatype and pattern which also use this logic.