Add new pattern: `tier`
`tier` is similar to the existing `value-partition` pattern but with the ability
to optionalize most of the axioms added. `value-partition` has been recoded to
use `tier`.

The main motivation for this change is that a value-partition is a solution to
a single problem -- the ontological representation of a continuous range; in
practice it was being used much more generally than this to introduce a
small hierarchy. The name `tier` better reflects this.