Add an :iri keyword to defentity
All interning owl entity forms can now take an :iri keyword. This was
previously supported with the non-interning forms, but was
syntactically impossible with the interning forms. Extended defentity
to enable this adds the capability to all the def forms. In addition,
we modifiy and rename extract-ontology-frame, so we can reuse the
functionality.