Move lookup to cache as we go
Previously lookup trawled through all vars, caching the results, then
removing cache when a new term was added; this had terrible
performance as well as tortuous code. Other methods now use this
cache. One test in memorize was updated as a result of this change;
the test appears buggy in its original version.