Add support for closed gems
Closed gems include a covering axiom. As they require all facets
values to be from the same facet, they have been made a different
function rather than an option on the existing gem function.

Additionally, this patch addresses a previously unknown issue with the
gem function, which was incorrectly handling the default ontology.